There's currently an ongoing debate off channel via email and Signal with daf regarding how to proceed with making muforth both more consistent between targets and have more of an actually convivial and humane interface.
The argument from this camp to refrain from being clever, and be obvious,
particularly when it comes to file and word naming.  There is no difference
in these files (except irq-task-experiments.mu4, that's just a demo that you
can use whatever for the output, cf the main repo) -- these are name changes
only to make things more obvious for the user.

Some found it particularly irritating that in the ARM directory the equates
files were clearly named as such, while in the MSP430 directory they were
hidden in a file that didn't have the word "equates" in the name.

If you want to use the original flavor, both will load (build vs build-v2).

	new file:   mu/target/MSP430/build-v2.mu4
	new file:   mu/target/MSP430/device/f5529-vectors-equates.mu4
	new file:   mu/target/MSP430/device/fr6989-vectors-equates.mu4
	new file:   mu/target/MSP430/device/g2553-vectors-equates.mu4
	new file:   mu/target/MSP430/device/supported-vectors-equates.mu4
	modified:   mu/target/MSP430/irq-task-experiments.mu4