Update the efm8 README's with a brief reminder to make sure you're a member of the group that controls serial-target.
NB: daf is making some potentially breaking changes to how memory is handled
on the 8051/AVR/PIC ... currently those changes are *not* incorporated here
until we've had time to test them. Reminder that if you want to play along on
the main repo at the very least you'll likely have to reflash; first test
this morning of the bb1 of his code failed to connect.

His primary interest right now is in getting the PIC to talk to the free
world; given how things have evolved in the past, we're going to hold fire
on updating locally and wait until the dust settles before melding his
code into this fork.  For now, if you want the bb1 to work, use this repo.

	modified:   mu/target/8051/silabs/efm8/README.md
	modified:   mu/target/8051/silabs/efm8/README.txt