Patch to get the stm32f3-discovery board
   A) actually communicating w/ the host
   B) actually working on OpenBSD

REMINDER: you need to pay attention to which ugenX.XX you're using
and you either need to change ownership of the specific device.endppoint
you're connecting via to $USER or chmod 660 on OpenBSD.

Yes, this is hacky and silly, no, noone here is going to write a
driver for it so don't ask. Remember to change it back.

	modified:   mu/target/ARM/board/stm32f3-discovery.mu4
	modified:   mu/target/ARM/debug/stlink-v2.mu4