AVR061: STK500 Communication Protocol
We needed this for figuring out the bootloader.

The two commands that matter: program page and read page.

"The mystery was: if the byte after the length is *not* an 'E', what is
it? optiboot ignores it entirely, the old bootloader (which was 4x
bigger) checks for E (eeprom) and assumes flash otherwise."

As an aside, I might as well mention here that avrdude

  * failed to read the fuse or lock bits
  * failed to recognize the boot section of flash
  * other failures to irritating to mention

Fortunately, using the method previously described with muforth and an
stm32f303 discovery board we were able to get a perfect image copy and
start taking it apart with, well, we checked with everything.  hexdump,
muforth, radare2.

Might as well mention: 99.99% of the Arduino code out there is painfully
ugly crap.  What the actual fsck.

This industry needs serious help.

	new file:   mu/target/AVR/ATmega328P/doc/doc2525.pdf