Add two updated files that don't break anything. ;)
We were able to recover the chip, due to the timing corrections in these
files and some gentle poking in jtag mode.

Alas ... the other clever new code on the main branch is still b0rk3d
dewds, so continue to use commit b7ed00

Also: a heartbeat is enabled OOTB if you follow the aforementioned v1
procedure.  500-1k resistor in front of LED straddling GND off PIN14
should start pulsing immediately after a successful prog (note that
verify will pull RESET and you'll lose your heartbeat but it comes back
on verification completion.)

So hurray, we're chatting with the target, albeit everything is still in
a futzy state.

	new file:   mu/target/ARM/stm32/clock.mu4
	modified:   mu/target/ARM/stm32/f3/03_c-init.mu4